move SelectivityHolders
