+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/AbstractScanCostEstimator.java
@@ -40,11 +40,10 @@ import com.splicemachine.db.iapi.sql.compile.Optimizer;
 import com.splicemachine.db.iapi.sql.conn.LanguageConnectionContext;
 import com.splicemachine.db.iapi.sql.dictionary.ColumnDescriptor;
 import com.splicemachine.db.iapi.sql.dictionary.ConglomerateDescriptor;
-import com.splicemachine.db.iapi.store.access.Qualifier;
 import com.splicemachine.db.iapi.store.access.StoreCostController;
 import com.splicemachine.db.iapi.types.DataValueDescriptor;
+import com.splicemachine.db.impl.sql.compile.selectivity.*;
 import org.apache.log4j.Logger;
-import scala.reflect.internal.Trees;
 
 import java.util.*;
