+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/selectivity/AbstractSelectivityHolder.java
@@ -29,7 +29,10 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
+package com.splicemachine.db.impl.sql.compile.selectivity;
+
+import com.splicemachine.db.impl.sql.compile.Predicate;
+import com.splicemachine.db.impl.sql.compile.QualifierPhase;
 
 /**
  *
