+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/selectivity/RangeSelectivity.java
@@ -29,11 +29,13 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
+package com.splicemachine.db.impl.sql.compile.selectivity;
 
 import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.db.iapi.store.access.StoreCostController;
 import com.splicemachine.db.iapi.types.DataValueDescriptor;
+import com.splicemachine.db.impl.sql.compile.Predicate;
+import com.splicemachine.db.impl.sql.compile.QualifierPhase;
 
 /**
  *
@@ -49,7 +51,7 @@ public class RangeSelectivity extends AbstractSelectivityHolder {
     private final double selectivityFactor;
     private final boolean useExtrapolation;
 
-    public RangeSelectivity(StoreCostController storeCost,DataValueDescriptor start, DataValueDescriptor stop,boolean includeStart, boolean includeStop,
+    public RangeSelectivity(StoreCostController storeCost, DataValueDescriptor start, DataValueDescriptor stop, boolean includeStart, boolean includeStop,
                             boolean fromExprIndex, int colNum, QualifierPhase phase, double selectivityFactor, boolean useExtrapolation, Predicate pred){
         super(fromExprIndex, colNum, phase, pred);
         this.start = start;
