+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/compile/costing/V1MergeSortJoinCostEstimation.java
@@ -18,7 +18,7 @@ import com.splicemachine.access.api.SConfiguration;
 import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.db.iapi.sql.compile.*;
 import com.splicemachine.db.iapi.sql.dictionary.ConglomerateDescriptor;
-import com.splicemachine.db.impl.sql.compile.SelectivityUtil;
+import com.splicemachine.db.impl.sql.compile.selectivity.SelectivityUtil;
 
 public class V1MergeSortJoinCostEstimation implements StrategyJoinCostEstimation {
     @Override
